refactor: rename predicate-utils to expandPredicate
