+++ packages/common/src/search/_internal/expandPredicate.ts
