+++ packages/common/src/search/_internal/portalSearchUsers.ts
@@ -15,7 +15,7 @@ import {
   IQuery,
 } from "../types";
 import { getNextFunction } from "../utils";
-import { expandPredicate } from "./ipredicate-utils";
+import { expandPredicate } from "./expandPredicate";
 
 /**
  * @private
